test: Add integration tests class, config
This adds config class for my integration tests.
