+++ pom.xml
@@ -111,7 +111,6 @@
         <dependency>
             <groupId>org.testcontainers</groupId>
             <artifactId>postgresql</artifactId>
-            <version>1.19.7</version>
             <scope>test</scope>
         </dependency>
