+++ src/main/java/dev/jsojka/basic_ecommerce_shop/testConfig/SecurityConfiguration.java
@@ -1,10 +1,9 @@
-package dev.jsojka.basic_ecommerce_shop.config;
+package dev.jsojka.basic_ecommerce_shop.testConfig;
 
 import dev.jsojka.basic_ecommerce_shop.auth.RateLimiterFilter;
 import dev.jsojka.basic_ecommerce_shop.users.UserDetailsServiceImpl;
 import dev.jsojka.basic_ecommerce_shop.logging.SessionDebugFilter;
 import dev.jsojka.basic_ecommerce_shop.users.IUserRepository;
-import dev.jsojka.basic_ecommerce_shop.users.IUserService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
@@ -14,7 +13,6 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
