+++ src/test/java/dev/jsojka/basic_ecommerce_shop/products/ProductRepositoryImplTest.java
@@ -1,13 +1,13 @@
 package dev.jsojka.basic_ecommerce_shop.products;
 
+import dev.jsojka.basic_ecommerce_shop.testConfig.AbstractIntegrationTest;
 import org.springframework.boot.jdbc.EmbeddedDatabaseConnection;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.test.context.ActiveProfiles;
 
 @DataJpaTest
-@AutoConfigureTestDatabase(connection = EmbeddedDatabaseConnection.H2)
 @ActiveProfiles("test")
-class ProductRepositoryImplTest {
+class ProductRepositoryImplTest extends AbstractIntegrationTest {
 
 }
