+++ src/test/java/dev/jsojka/basic_ecommerce_shop/testConfig/AbstractIntegrationTest.java
@@ -1,0 +1,33 @@
+package dev.jsojka.basic_ecommerce_shop.testConfig;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
+import org.springframework.test.context.DynamicPropertyRegistry;
+import org.springframework.test.context.DynamicPropertySource;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+
+import static dev.jsojka.basic_ecommerce_shop.testConfig.TestConfig.postgres;
+
+//@ExtendWith(SpringExtension.class)
+//@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@Import(TestConfig.class)
+public abstract class AbstractIntegrationTest {
+
+    @DynamicPropertySource
+    static void registerPgProperties(DynamicPropertyRegistry registry) {
+        registry.add("spring.datasource.url", postgres::getJdbcUrl);
+        registry.add("spring.datasource.username", postgres::getUsername);
+        registry.add("spring.datasource.password", postgres::getPassword);
+        registry.add("spring.jpa.properties.hibernate.dialect", () -> "org.hibernate.dialect.PostgreSQLDialect");
+    }
+
+    @Test
+    void connectionEstablished() {
+        Assertions.assertThat(postgres.isCreated()).isTrue();
+        Assertions.assertThat(postgres.isRunning()).isTrue();
+    }
+}
